Design-change (Grid-Element)
- outsourcing media component for managing more than two cases of media-container in cards
- fix several style issues after that change
- clean up files
