+++ src/app/components/grid-element/entity-media-container/entity-media-container.component.html
@@ -1,0 +1,25 @@
+@if (element() | isEntity) {
+  @if (videoPreview(); as videoPreviewUrl) {
+    <div class="image cursor-pointer" [style.--bg-color]="backgroundColor()">
+      <video
+        #videoEl
+        [src]="videoPreviewUrl"
+        loop
+        preload
+        style="width: 100%; object-fit: cover"
+        (mousemove)="updateVideoPreview(videoEl, $event)"
+        (mouseleave)="resetVideoPreview(videoEl)"
+      ></video>
+    </div>
+  } @else {
+    <div class="image cursor-pointer" [style.--bg-color]="backgroundColor()">
+      <anim-img
+        [routerLink]="link() ? ['/entity', element()._id] : undefined"
+        [src]="imageSource()"
+        [alt]="element().name"
+        matTooltip="{{ 'Click to view object' | translate }}"
+        [matTooltipShowDelay]="1000"
+      />
+    </div>
+  }
+}
