+++ src/app/components/grid-element/grid-element.component.html
@@ -1,111 +1,10 @@
 @if (element(); as element) {
   <!-- Default behaviour: Navigate to element when clicked -->
-  @if (!disableNavigationOnClick()) {
-    @if (element | isEntity) {
-      @if (videoPreview(); as videoPreviewUrl) {
-        <video
-          #videoEl
-          [src]="videoPreviewUrl"
-          loop
-          preload
-          style="width: 100%; object-fit: cover"
-          class="image cursor-pointer"
-          (mousemove)="updateVideoPreview(videoEl, $event)"
-          (mouseleave)="resetVideoPreview(videoEl)"
-        ></video>
-      } @else {
-        <anim-img
-          class="image cursor-pointer"
-          [routerLink]="['/entity', element._id]"
-          matTooltip="{{ 'Click to view object' | translate }}"
-          [matTooltipShowDelay]="1000"
-          [src]="imageSource()"
-          [style.--bg-color]="backgroundColor()"
-          [alt]="element.name"
-        />
-      }
-    } @else {
-      <div
-        class="images cursor-pointer"
-        [style.--bg-color]="backgroundColor()"
-        [class.single-image]="imageSources.length == 1"
-        [routerLink]="['/compilation', element._id]"
-        matTooltip="{{ 'Click to view collection' | translate }}"
-        [matTooltipShowDelay]="1000"
-      >
-        @for (source of imageSources(); track source) {
-          <anim-img [alt]="element.name" [src]="source" />
-        }
-      </div>
-    }
+  @if (element | isEntity) {
+    <app-entity-media-container [element]="element" [link]="!disableNavigationOnClick()" />
   } @else {
-    @if (element | isEntity) {
-      <anim-img
-        class="image"
-        [style.--bg-color]="backgroundColor()"
-        [src]="imageSource()"
-        alt="{{ element.name }}"
-        [class.private]="!element.online"
-      />
-    } @else {
-      <div
-        class="images"
-        [style.--bg-color]="backgroundColor()"
-        [class.single-image]="imageSources().length == 1"
-      >
-        @for (source of imageSources(); track source) {
-          <anim-img alt="{{ element.name }}" [src]="source" />
-        }
-      </div>
-    }
+    <app-collection-media-container [element]="element" [link]="!disableNavigationOnClick()" />
   }
-  <!-- @if (!disableTypeInfo()) {
-    <div class="entity-type">
-      <mat-icon
-        (click)="openExploreDialog(element)"
-        class="button cursor-pointer"
-        [matTooltip]="((element | isEntity) ? 'Preview object' : 'Preview collection') | translate"
-        >open_in_browser</mat-icon
-      >
-      @if (tooltipContent.length > 20) {
-        <mat-icon class="help cursor-default" [matTooltip]="tooltipContent()">description</mat-icon>
-      }
-      @if (element | isCompilation) {
-        @if (isRecentlyAnnotated()) {
-          <mat-icon
-            class="help cursor-default"
-            matTooltip="{{ 'This collection has recently been annotated' | translate }}"
-            >access_time</mat-icon
-          >
-        }
-        @if (isPasswordProtected()) {
-          <mat-icon class="help cursor-default" matTooltip="{{ 'Password protected' | translate }}"
-            >lock</mat-icon
-          >
-        }
-        <mat-icon class="cursor-default" [matTooltip]="collectionQuantityText()">{{
-          collectionQuantityIcon()
-        }}</mat-icon>
-        <mat-icon class="help cursor-default" matTooltip="{{ 'Collection' | translate }}">{{
-          icons.collection
-        }}</mat-icon>
-      }
-      @if (element | isEntity) {
-        @if (quickAddToCollectionMenu(); as menu) {
-          <mat-icon
-            class="button cursor-pointer"
-            matTooltip="{{ 'Quick add to collection' | translate }}"
-            [matMenuTriggerFor]="menu"
-            (click)="selectObject(element._id)"
-            >playlist_add</mat-icon
-          >
-        }
-        <mat-icon class="help cursor-default" [matTooltip]="helpTooltip()">{{
-          helpIcon()
-        }}</mat-icon>
-      }
-    </div>
-  } -->
   <div class="title" [class.private]="isPrivate()">
     <div class="element-name">
       {{ element.name }}
@@ -118,7 +17,7 @@
     }
     @if (!(element | isEntity)) {
       <div class="annotation-counter">
-        <mat-icon class="material-symbols-filled card-icon">deployed_code</mat-icon>
+        <mat-icon class="material-symbols-filled">deployed_code</mat-icon>
         <span>{{ entitiesCount() }}</span>
       </div>
     }
