+++ src/app/pages/profile-page/compilations/compilations.component.scss
@@ -31,15 +31,6 @@
   }
 }
 
-.grid-item {
-  position: relative;
-  cursor: context-menu;
-
-  .actionbutton {
-    @include mat.elevation(1);
-  }
-}
-
 .group-grid-item {
   @include mat.elevation(2);
 
@@ -57,28 +48,6 @@
 
 .detail-block {
   padding: 10px 0;
-}
-
-.card-icon {
-  position: absolute !important;
-  top: 15px;
-  left: 15px;
-  color: #f5f5f6;
-}
-
-.actionbutton {
-  display: grid;
-  place-items: center;
-  width: 24px;
-  height: 24px;
-  line-height: 24px;
-  padding: 0;
-  .material-icons {
-    font-size: 16px;
-    width: 16px;
-    height: 16px;
-    line-height: 16px;
-  }
 }
 
 #profile-page-content {
